livekit-agent/pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "LiveKit Voice Agent"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-silero>=0.7.0",
    "livekit-plugins-assemblyai>=0.2.0",
    "livekit-plugins-groq>=0.2.0",
    "livekit-plugins-cartesia>=0.4.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = []
